+++ itbl.c
@@ -1,6 +1,6 @@
 #include "asm6502.h"
 
-int am_size[13] = {1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 2, 2};
+u16 am_size[13] = {1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 2, 2};
 
 idesc itbl[56] = {
 	{"ADC", {INV, INV, 0x69, INV, 0x65, 0x75, INV, 0x6d, 0x7d, 0x79, INV, 0x61, 0x71}},
